住院费用明细获取修正

+++ FE.Handle.Request/HosOrderHandle.cs
@@ -48,7 +48,7 @@ namespace FE.Handle.Request
             }
         }
 
-        public List<HosOrderItem> GetHosOrdersItems(int zyh)
+        public List<HosOrderItem> GetHosOrdersItems(decimal zyh)
         {
             try
             {
